move dto fixtures to plugins
