+++ tests/conftest.py
@@ -123,6 +123,7 @@ pytest_plugins = [
     "tests.fixture_plugins.orders_fixtures.order_to_submit_fixtures",
     "tests.fixture_plugins.orders_fixtures.status_data_fixtures",
     "tests.fixture_plugins.pacbio_fixtures.context_fixtures",
+    "tests.fixture_plugins.pacbio_fixtures.dto_fixtures",
     "tests.fixture_plugins.pacbio_fixtures.file_data_fixtures",
     "tests.fixture_plugins.pacbio_fixtures.metrics_fixtures",
     "tests.fixture_plugins.pacbio_fixtures.name_fixtures",
